refactor(input-error-message): extract field name suggestions to own module
